move email to models
